+++ CHANGELOG.md
@@ -4,6 +4,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.3.0] - 2016-04-28
+### Add
+- String.camel_case
+- String.kebab_case
+- String.downcase_first
+- String.upcase_first
+- String.snake_case
+- String.words
+- Collection.find_last
+- Collection.pfind_last
+
 ## [0.2.0] - 2016-04-27
 ### Update
 - documentation for pmap and pfilter
@@ -25,5 +36,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - ExCheck
 
 [Unreleased]: https://github.com/TFarla/exdash/compare/master...HEAD
-[0.2.0]: https://github.com/TFarla/exdash/compare/0.2.0...HEAD
+[0.3.0]: https://github.com/TFarla/exdash/compare/0.3.0...HEAD
+[0.2.0]: https://github.com/TFarla/exdash/compare/0.2.0...0.3.0
 [0.0.1]: https://github.com/TFarla/exdash/compare/0.0.1...0.2.0